+++ tool/junit_user/result/result.go
@@ -109,8 +109,8 @@ func (this *Result) Success() bool {
 //ChartVals
 func (this *Result) ChartVals() []*tool.ChartVal {
 	return []*tool.ChartVal{
-		&tool.ChartVal{"Errors", this.Report.Errors, true, this.FileId},
-		&tool.ChartVal{"Failures", this.Report.Failures, true, this.FileId},
+		&tool.ChartVal{"Errors", float64(this.Report.Errors), true, this.FileId},
+		&tool.ChartVal{"Failures", float64(this.Report.Failures), true, this.FileId},
 	}
 }
